RabbitMQ producer 부분 수정
