+++ collector/build.gradle
@@ -1,3 +1,4 @@
 dependencies {
+    compile project(":domain")
     compile("org.springframework.boot:spring-boot-starter-amqp")
 }
