+++ collector/src/main/java/com/sippet/collector/Producer.java
@@ -1,24 +1,72 @@
 package com.sippet.collector;
 
+import com.sippet.domain.util.NullChecker;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.amqp.core.*;
+import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.amqp.support.converter.SimpleMessageConverter;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 
-@Component
+@Configuration
 public class Producer {
     private static Logger log = LoggerFactory.getLogger(Producer.class);
 
-    @Autowired
-    private RabbitTemplate rabbitTemplate;
+//    @Autowired
+//    private RabbitTemplate rabbitTemplate;
 
-    public Producer(RabbitTemplate rabbitTemplate) {
-        this.rabbitTemplate = rabbitTemplate;
-    }
+//    public Producer(RabbitTemplate rabbitTemplate) {
+//        this.rabbitTemplate = rabbitTemplate;
+//    }
+
+//    final static String queueName = "testMQ";
+//
+//    private CachingConnectionFactory cachingConnectionFactory;
+//
+//    @Bean
+//    ConnectionFactory connectionFactory() {
+//        //return new CachingConnectionFactory("localhost", 5672);
+//        if(NullChecker.check(cachingConnectionFactory)) {
+//            cachingConnectionFactory = new CachingConnectionFactory("localhost", 5672);
+//        }
+//
+//        return cachingConnectionFactory;
+//    }
+//
+//    @Bean
+//    Queue queue() {
+//        return new Queue(queueName, false);
+//    }
+//
+//    @Bean
+//    TopicExchange exchange() {
+//        return new TopicExchange("spring-boot-exchange");
+//    }
+//
+//    @Bean
+//    Binding binding(Queue queue, TopicExchange topicExchange) {
+//        log.info("check exchange" + topicExchange.getName());
+//        return BindingBuilder.bind(queue).to(topicExchange).with(queueName);
+//    }
+//
+//    @Bean
+//    AmqpTemplate amqpTemplate() {
+//        RabbitTemplate template = new RabbitTemplate();
+//        template.setConnectionFactory(connectionFactory());
+//        template.setRoutingKey("testMQ");
+//        //template.setQueue(queue());
+//        template.setMessageConverter(new SimpleMessageConverter());
+//        return template;
+//    }
 
     public void sendTo(String routingKey, String message) {
         log.info("전송>>...");
-        this.rabbitTemplate.convertAndSend(routingKey, message);
+        RabbitMQConfiguration.amqpTemplate().convertAndSend(routingKey, message);
+        //this.rabbitTemplate.convertAndSend(routingKey, message);
     }
 }
